// src/app/join/schema.ts
import { z } from "zod";


export const AccountCreationSchema = z.object({
    firstName: z.string().min(2, "First name must be at least 2 characters"),
    lastName: z.string().min(2, "Last name must be at least 2 characters"),
    email: z.string().email("Invalid email address"),
    phone: z.string().min(10, "Phone number must be at least 10 characters"),
    // password: z.string().min(8, "Password must be at least 8 characters"),
    // confirmPassword: z.string().min(8, "Password must be at least 8 characters"),
});